fix: Update Cell to Nib
