+++ cozy/cozy/Source/ViewControllers/Map/Cell/MapSelectCell.swift
