+++ cozy/cozy/Source/ViewControllers/Map/VC/MapVC.swift
@@ -11,12 +11,15 @@ import UIKit
 class MapVC: UIViewController {
 
     private let mapIdentifier1: String = "mapSelectCell"
-    private let mapIdentifier2: String = "mapListCell"
+    private let mapIdentifier2: String = "bookListCell"
 
     @IBOutlet weak var mapTableView: UITableView!
 
     override func viewDidLoad() {
         super.viewDidLoad()
+
+        let nibName = UINib(nibName: "BookListCell", bundle: nil)
+        mapTableView.register(nibName, forCellReuseIdentifier: mapIdentifier2)
 
         mapTableView.delegate = self
         mapTableView.dataSource = self
@@ -51,17 +54,20 @@ extension MapVC: UITableViewDelegate, UITableViewDataSource {
             let cell = tableView.dequeueReusableCell(withIdentifier: mapIdentifier1) as! MapSelectCell
             return cell
         } else {
-            let cell = tableView.dequeueReusableCell(withIdentifier: mapIdentifier2) as! MapListCell
-
-            cell.bookstoreImageView.image = UIImage(named: "asdfdghfgjhj")
-            cell.nameLabel.text = "코지서점"
-            cell.addressLabel.text = "서울특별시 용산구 한강대로 10길"
-
-            cell.tag1.setTitle("    #베이커리    ", for: .normal)
-            cell.tag2.setTitle("    #심야책방    ", for: .normal)
-            cell.tag3.setTitle("    #맥주    ", for: .normal)
-
+            let cell = tableView.dequeueReusableCell(withIdentifier: mapIdentifier2) as! BookListCell
             return cell
+
+//            let cell = tableView.dequeueReusableCell(withIdentifier: mapIdentifier2) as! MapListCell
+//
+//            cell.bookstoreImageView.image = UIImage(named: "asdfdghfgjhj")
+//            cell.nameLabel.text = "코지서점"
+//            cell.addressLabel.text = "서울특별시 용산구 한강대로 10길"
+//
+//            cell.tag1.setTitle("    #베이커리    ", for: .normal)
+//            cell.tag2.setTitle("    #심야책방    ", for: .normal)
+//            cell.tag3.setTitle("    #맥주    ", for: .normal)
+
+//            return cell
         }
     }
